requirements.txt
# Core Data Processing
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0

# MongoDB (No SQL databases)
pymongo==4.5.0

# Scheduling & Automation
schedule==1.2.0
apscheduler==3.10.4

# ML & NLP
scikit-learn==1.3.0
xgboost==2.0.0
spacy==3.6.1
transformers==4.33.0
torch==2.0.1

# LLM Integration
openai==0.28.0
langchain==0.0.300

# API Development
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Resume Parsing
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.10.2

# Data Visualization
plotly==5.17.0
streamlit==1.27.0

# Job Market APIs
requests==2.31.0
beautifulsoup4==4.12.2

# Feature Engineering
category-encoders==2.6.2
imbalanced-learn==0.11.0

# Model Persistence
joblib==1.3.2
pickle5==0.0.12

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Workflow Orchestration (Optional)
# apache-airflow==2.7.1  # Removed - using schedule instead
